fix: circular denpendency cause CI fail
